requirements.txt
# Core MLOps Dependencies
boto3==1.34.0
botocore==1.34.0
sagemaker==2.199.0
awscli==1.32.0

# ML and Data Processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
joblib==1.3.2

# Deep Learning
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector Database and RAG
chromadb==0.4.18
langchain==0.0.350
langchain-aws==0.0.4

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Message Queue
celery==5.3.4
redis==5.0.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
moto==4.2.14
localstack==3.0.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# AWS CDK (Infrastructure as Code)
aws-cdk-lib==2.105.0
constructs==10.3.0 